refactor: useProductList, useProductDetail 커스텀훅 분리
