+++ src/pages/ProductListPage/index.jsx
@@ -1,23 +1,12 @@
-import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import React from "react";
 import styled from "styled-components";
-
-import { getProductList } from "../../store/actions";
 
 import Spinner from "../../components/common/Spinner";
 import ProductCard from "./ProductCard";
+import { useProductList } from "../../hooks/useProductList";
 
 function ProductListPage() {
-  const dispatch = useDispatch();
-  const {
-    data: productList,
-    loading: isLoading,
-    errorMessage,
-  } = useSelector((state) => state.productListReducer.productList);
-
-  useEffect(() => {
-    dispatch(getProductList());
-  }, []);
+  const { productList, isLoading, errorMessage } = useProductList();
 
   if (isLoading) return <Spinner />;
   if (errorMessage) return <div>😱 Error: {errorMessage} 😱</div>;
